Massive automated class cleanup

+++ src/main/java/tragicneko/tragicmc/client/render/RenderSoulChest.java
@@ -1,11 +1,5 @@
 package tragicneko.tragicmc.client.render;
 
-import cpw.mods.fml.common.FMLLog;
-import cpw.mods.fml.relauncher.Side;
-import cpw.mods.fml.relauncher.SideOnly;
-
-import java.util.Calendar;
-
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockChest;
 import net.minecraft.client.model.ModelChest;
@@ -17,6 +11,10 @@ import net.minecraft.util.ResourceLocation;
 
 import org.lwjgl.opengl.GL11;
 import org.lwjgl.opengl.GL12;
+
+import cpw.mods.fml.common.FMLLog;
+import cpw.mods.fml.relauncher.Side;
+import cpw.mods.fml.relauncher.SideOnly;
 
 @SideOnly(Side.CLIENT)
 public class RenderSoulChest extends TileEntitySpecialRenderer
@@ -108,7 +106,7 @@ public class RenderSoulChest extends TileEntitySpecialRenderer
 				GL11.glTranslatef(0.0F, 0.0F, -1.0F);
 			}
 
-			GL11.glRotatef((float)short1, 0.0F, 1.0F, 0.0F);
+			GL11.glRotatef(short1, 0.0F, 1.0F, 0.0F);
 			GL11.glTranslatef(-0.5F, -0.5F, -0.5F);
 			float f1 = te.prevLidAngle + (te.lidAngle - te.prevLidAngle) * partialTick;
 			float f2;
